refactor: editor composable 함수로 이름 변경
